Climbing State and Last Key Aiming
Added a climbing state to the player, but I haven't tested it or coded in anything that triggers it. Also updated the shooting controls to aim based on the 2 arrow keys most recently pressed. The aiming currently behaves weirdly if you hold down more than 2 arrow keys at once, but this will be fixed soon.