Player Dictionary and Movement Updates
Started working on the dash. Added a dictionary that stores each upgrade and whether each player has taken it or not (I made this an IntDict to facilitate the potential for a skill tree with multiple different "levels" in the same upgrade). Also tweaked some numbers in the base player movement and shooting to get a better fit for the game being full screen.