player_phase_6
Enemy and Enemy type classes are created to implement the Type Object pattern. An enemy factory class is used to create different types of enemies dynamically. Apple and Banana, as the various types of enemies, extend the enemy class. The Playhandler class is used to control the play flow and mimic the game loop pattern. Different player states are extracted to different types of state classes to implement the PlayState interface. All types of state instances are preloaded into a hashmap.