Platform Overhaul
Changed the code for platforms so that the interaction occurs in the player's code, rather than the platform's, so that it interacts properly with the player's state machine. This introduces a problem with climbing that I will fix in my next commit. I also temporarily commented out enemy spawning because it is currently bugged, but also so I could test my movement changes more easily.